xml/parsing/manakai/entities-2.dat

#data
<!DOCTYPE a[
<!ENTITY%>
]><a/>
#errors
1;1;no XML decl;;;s
2;10;no space between params;;;m
2;10;no md name;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities

#data
<!DOCTYPE a[
<!ENTITY%
#errors
1;1;no XML decl;;;s
2;10;parser:EOF;;;m
2;10;parser:EOF;;;m
2;10;parser:EOF;;;m
2;10;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY %>
]><a/>
#errors
1;1;no XML decl;;;s
2;11;no md name;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities

#data
<!DOCTYPE a[
<!ENTITY %
#errors
1;1;no XML decl;;;s
2;11;parser:EOF;;;m
2;11;parser:EOF;;;m
2;11;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE x [<!ENTITY%=>]><x/>
#errors
1;1;no XML decl;;;s
1;23;no space between params;;;m
1;24;no md def;;;m
#document
| <!DOCTYPE x>
| <{} x>

#data
<!DOCTYPE x [<!ENTITY%<>]><x/>
#errors
1;1;no XML decl;;;s
1;23;no space between params;;;m
1;24;no md def;;;m
#document
| <!DOCTYPE x>
| <{} x>

#data
<!DOCTYPE x [<!ENTITY%">]><x/>
#errors
1;1;no XML decl;;;s
1;23;no space between params;;;m
1;24;no md def;;;m
#document
| <!DOCTYPE x>
| <{} x>

#data
<!DOCTYPE x [<!ENTITY%'>]><x/>
#errors
1;1;no XML decl;;;s
1;23;no space between params;;;m
1;24;no md def;;;m
#document
| <!DOCTYPE x>
| <{} x>

#data
<!DOCTYPE x [<!ENTITY% >]><x/>
#errors
1;1;no XML decl;;;s
1;23;no space between params;;;m
1;24;no md name;;;m
#document
| <!DOCTYPE x>
| <{} x>

#data
<!DOCTYPE x [<!ENTITY%%>]><x/>
#errors
1;1;no XML decl;;;s
1;23;no space between params;;;m
1;24;no md def;;;m
#document
| <!DOCTYPE x>
| <{} x>

#data
<!DOCTYPE x [<!ENTITY%`>]><x/>
#errors
1;1;no XML decl;;;s
1;23;no space between params;;;m
1;24;no md def;;;m
#document
| <!DOCTYPE x>
| <{} x>

#data
<!DOCTYPE x [<!ENTITY %=>]><x/>
#errors
1;1;no XML decl;;;s
1;25;no md def;;;m
#document
| <!DOCTYPE x>
| <{} x>

#data
<!DOCTYPE x [<!ENTITY %<>]><x/>
#errors
1;1;no XML decl;;;s
1;25;no md def;;;m
#document
| <!DOCTYPE x>
| <{} x>

#data
<!DOCTYPE x [<!ENTITY %#>]><x/>
#errors
1;1;no XML decl;;;s
1;24;no space between params;;;m
#document
| <!DOCTYPE x>
| <{} x>

#data escaped
<!DOCTYPE a[
<!ENTITY % abc  \u0009
#errors
1;1;no XML decl;;;s
2;18;parser:EOF;;;m
2;18;parser:EOF;;;m
2;18;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc  #
#errors
1;1;no XML decl;;;s
2;17;xml:dtd:string after name;;;m
2;18;parser:EOF;;;m
2;18;parser:EOF;;;m
2;18;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc PUBLIC
#errors
1;1;no XML decl;;;s
2;22;parser:EOF;;;m
2;22;parser:EOF;;;m
2;22;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data escaped
<!DOCTYPE a[
<!ENTITY % abc PUBLIC\u0020
#errors
1;1;no XML decl;;;s
2;23;parser:EOF;;;m
2;23;parser:EOF;;;m
2;23;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc PUBLIC #
#errors
1;1;no XML decl;;;s
2;23;no DOCTYPE literal;;;m
2;24;parser:EOF;;;m
2;24;parser:EOF;;;m
2;24;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc SYSTEM
#errors
1;1;no XML decl;;;s
2;22;parser:EOF;;;m
2;22;parser:EOF;;;m
2;22;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data escaped
<!DOCTYPE a[
<!ENTITY % abc SYSTEM\u0020
#errors
1;1;no XML decl;;;s
2;23;parser:EOF;;;m
2;23;parser:EOF;;;m
2;23;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc PUBLIC #
#errors
1;1;no XML decl;;;s
2;23;no DOCTYPE literal;;;m
2;24;parser:EOF;;;m
2;24;parser:EOF;;;m
2;24;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc PUBLIC "a"x
#errors
1;1;no XML decl;;;s
2;26;no DOCTYPE literal;;;m
2;27;parser:EOF;;;m
2;27;parser:EOF;;;m
2;27;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc PUBLIC "a" x
#errors
1;1;no XML decl;;;s
2;27;no DOCTYPE literal;;;m
2;28;parser:EOF;;;m
2;28;parser:EOF;;;m
2;28;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc SYSTEM "a"x
#errors
1;1;no XML decl;;;s
2;26;no space between params;;;m
2;26;string after SYSTEM literal;;;m
2;27;parser:EOF;;;m
2;27;parser:EOF;;;m
2;27;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc SYSTEM "a" x
#errors
1;1;no XML decl;;;s
2;27;string after SYSTEM literal;;;m
2;28;parser:EOF;;;m
2;28;parser:EOF;;;m
2;28;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc PUBLIC "a" "b"x
#errors
1;1;no XML decl;;;s
2;30;no space between params;;;m
2;30;string after SYSTEM literal;;;m
2;31;parser:EOF;;;m
2;31;parser:EOF;;;m
2;31;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc PUBLIC "a" "b" x
#errors
1;1;no XML decl;;;s
2;31;string after SYSTEM literal;;;m
2;32;parser:EOF;;;m
2;32;parser:EOF;;;m
2;32;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC>
]><a/>
#errors
1;1;no XML decl;;;s
2;21;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC >
]><a/>
#errors
1;1;no XML decl;;;s
2;22;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % bc SYSTEM>
]><a/>
#errors
1;1;no XML decl;;;s
2;21;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % bc SYSTEM >
]><a/>
#errors
1;1;no XML decl;;;s
2;22;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC "x" "y">
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "x" "y" >

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC "x""y">
]><a/>
#errors
1;1;no XML decl;;;s
2;25;no space between params;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "x" "y" >

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC "x"'y'>
]><a/>
#errors
1;1;no XML decl;;;s
2;25;no space between params;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "x" "y" >

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC 'x'"y">
]><a/>
#errors
1;1;no XML decl;;;s
2;25;no space between params;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "x" "y" >

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC 'x''y'>
]><a/>
#errors
1;1;no XML decl;;;s
2;25;no space between params;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "x" "y" >

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC "x" "y" NDATA#
#errors
1;1;no XML decl;;;s
2;35;no space between params;;;m
2;36;no root element;;;m
2;36;parser:EOF;;;m
2;36;parser:EOF;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC "x" "y" NDATA #
#errors
1;1;no XML decl;;;s
2;37;no root element;;;m
2;37;parser:EOF;;;m
2;37;parser:EOF;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC "x" "y" NDATA z#
#errors
1;1;no XML decl;;;s
2;38;no root element;;;m
2;38;parser:EOF;;;m
2;38;parser:EOF;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC "x" "y" NDATA z #
#errors
1;1;no XML decl;;;s
2;38;string after md def;;;m
2;39;no root element;;;m
2;39;parser:EOF;;;m
2;39;parser:EOF;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC "x" "y" NDATA z>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:param entity with ndata;;%bc;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % bc "" "x" "y" >

#data
<!DOCTYPE a[
<!ENTITY % b ""
#errors
1;1;no XML decl;;;s
2;16;no root element;;;m
2;16;parser:EOF;;;m
2;16;parser:EOF;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % b ''
#errors
1;1;no XML decl;;;s
2;16;no root element;;;m
2;16;parser:EOF;;;m
2;16;parser:EOF;;;m
#document
| <!DOCTYPE a>
#entities

#data escaped
<!DOCTYPE a[
<!ENTITY % b ""\u0020
#errors
1;1;no XML decl;;;s
2;17;no root element;;;m
2;17;parser:EOF;;;m
2;17;parser:EOF;;;m
#document
| <!DOCTYPE a>
#entities

#data escaped
<!DOCTYPE a[
<!ENTITY % b ''\u0020
#errors
1;1;no XML decl;;;s
2;17;no root element;;;m
2;17;parser:EOF;;;m
2;17;parser:EOF;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % b "">
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b ''>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b "xxxxx">
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "xxxxx" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b 'xxxxx'>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "xxxxx" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b "yyy" >
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "yyy" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b 'yyy' >
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "yyy" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b "a&#x24;b">
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "a$b" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b '&#32;'>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b " " "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b "&amp;abc">
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "&amp;abc" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b "c">
<!ENTITY % b "d">
]><a/>
#errors
1;1;no XML decl;;;s
3;1;duplicate entity decl;;%b;;w
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "c" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b "c">
<!ENTITY b "d">
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY b "d" "" "" >
<!ENTITY % b "c" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b SYSTEM "x" >
%c;
<!ENTITY % d SYSTEM "y" >
]><a/>
#errors
1;1;no XML decl;;;s
3;1;entity not declared;;%c;;m
3;1;stop processing;;;i
4;1;xml:dtd:entity ignored;;;w
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % b "" "" "x" >

#data
<?xml version="1.0" standalone="yes"?>
<!DOCTYPE a[
<!ENTITY % b SYSTEM "x" >
%c;
<!ENTITY % d SYSTEM "y" >
]><a/>
#errors
4;1;entity not declared;;%c;;m
#xml-standalone
true
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % b "" "" "x" >
| <!ENTITY % d "" "" "y" >
